+++ test/support/test_utils.ex
@@ -1,0 +1,10 @@
+defmodule Omedis.TestUtils do
+  @moduledoc false
+
+  def time_after(seconds_offset) do
+    DateTime.utc_now()
+    |> DateTime.add(seconds_offset)
+    |> DateTime.to_naive()
+    |> NaiveDateTime.truncate(:second)
+  end
+end
